requirements.txt
asgiref==3.5.2
Babel==2.10.3
Django==4.0.6
django-phone-field==1.8.1
django-phonenumber-field==6.3.0
django-ratelimit==3.0.1
phonenumbers==8.12.51
Pillow==9.2.0
psycopg2==2.9.3
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
